requirements.txt
anyio==3.6.2
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fastapi==0.89.1
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
idna==3.4
joblib==1.2.0
jsonpickle==3.0.1
kiwisolver==1.4.4
matplotlib==3.6.3
more-itertools==9.0.0
music21==8.1.0
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webcolors==1.12
websockets==10.4
